[FEAT] Back Game Beta
Adding all the game logic in the backend with queue system, sockets and connection to a game room. The game should be functional too, I just followed a tuto to make a single player pong and mirror it to add controller for the right player. New DB entity: Room

TODO: Next step is to test if the playground works in the front side.